pyproject.toml
[project]
name = "papercal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "icalendar>=6.3.1",
    "pillow>=11.2.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "ruff>=0.12.0",
]
